# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MMA Tournament API
  version: 1.0.0
  description: |
    API for managing MMA tournament data including athletes, events, rankings, and matches.
    Supports multiple tournaments (UFC, Lion Championship, etc.) with comprehensive CRUD operations.
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: http://localhost:4000/api
    description: Local development API base
...
paths:
...
  /api/athletes:
    get:
      summary: Get all athletes
      description: Retrieve all athletes for a tournament with optional filtering
      tags:
        - Athletes
      parameters:
        - name: tournament
          in: query
          description: Tournament identifier
          schema:
            type: string
            default: ufc
            enum: [ufc, lion]
        - name: division
          in: query
          description: Filter by division
          schema:
            type: string
            example: lightweight
        - name: gender
          in: query
          description: Filter by gender
          schema:
            type: string
            enum: [male, female]
        - name: search
          in: query
          description: Search athletes by name
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    type: string
                    example: ufc
                  count:
                    type: integer
                    example: 1
                  athletes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Athlete'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new athlete
      description: Add a new athlete to the database
      tags:
        - Athletes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AthleteInput'
      responses:
        '201':
          description: Athlete created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  athlete:
                    $ref: '#/components/schemas/Athlete'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/athletes/{id}:
    get:
      summary: Get athlete by ID
      description: Retrieve a specific athlete by their ID
      tags:
        - Athletes
      parameters:
        - $ref: '#/components/parameters/AthleteId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  athlete:
                    $ref: '#/components/schemas/Athlete'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update athlete
      description: Update an existing athlete's information
      tags:
        - Athletes
      parameters:
        - $ref: '#/components/parameters/AthleteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AthleteInput'
      responses:
        '200':
          description: Athlete updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  athlete:
                    $ref: '#/components/schemas/Athlete'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete athlete
      description: Remove an athlete from the database
      tags:
        - Athletes
      parameters:
        - $ref: '#/components/parameters/AthleteId'
      responses:
        '200':
          description: Athlete deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Athlete deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/events:
    get:
      summary: Get all events
      description: Retrieve all events for a tournament
      tags:
        - Events
      parameters:
        - name: tournament
          in: query
          description: Tournament identifier
          schema:
            type: string
            default: ufc
            enum: [ufc, lion]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    type: string
                    example: ufc
                  count:
                    type: integer
                    example: 0
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new event
      description: Add a new event to the database
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: '#/components/schemas/Event'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/events/{id}:
    get:
      summary: Get event by ID
      description: Retrieve a specific event with its matches
      tags:
        - Events
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: '#/components/schemas/Event'
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/Match'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update event
      description: Update an existing event's information
      tags:
        - Events
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  event:
                    $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete event
      description: Remove an event from the database
      tags:
        - Events
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/events/{id}/matches:
    post:
      summary: Add match to event
      description: Create a new match for a specific event
      tags:
        - Events
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchInput'
      responses:
        '201':
          description: Match created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  match:
                    $ref: '#/components/schemas/Match'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Event not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/promotions/{id}/events:
    post:
      summary: Add event to promotion
      description: Add an event ID to a promotion's events array
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
              properties:
                event_id:
                  type: string
                  example: ufc-310
      responses:
        '201':
          description: Event added to promotion successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event added to promotion successfully
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Promotion not found
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Remove event from promotion
      description: Remove an event ID from a promotion's events array
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
        - name: event_id
          in: query
          required: true
          description: Event ID to remove
          schema:
            type: string
            example: ufc-310
      responses:
        '200':
          description: Event removed from promotion successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event removed from promotion successfully
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '404':
          description: Promotion or event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Promotion not found
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Athlete:
      type: object
      required:
        - name
        - tournament
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Test Fighter
        tournament:
          type: string
          example: ufc
        division:
          type: string
          example: lightweight
        country:
          type: string
          example: USA
        flag:
          type: string
          nullable: true
          example: null
        gender:
          type: string
          enum: [male, female]
          nullable: true
          example: null
        wins:
          type: integer
          example: 10
        losses:
          type: integer
          example: 2
        draws:
          type: integer
          example: 0
        image_url:
          type: string
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
        updated_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
    AthleteInput:
      type: object
      required:
        - name
        - tournament
      properties:
        name:
          type: string
          example: Test Fighter
        tournament:
          type: string
          example: ufc
        division:
          type: string
          example: lightweight
        country:
          type: string
          example: USA
        flag:
          type: string
          example: 🇺🇸
        gender:
          type: string
          enum: [male, female]
          example: male
        wins:
          type: integer
          example: 10
        losses:
          type: integer
          example: 2
        draws:
          type: integer
          example: 0
        image_url:
          type: string
          example: https://example.com/fighter.jpg
    Event:
      type: object
      required:
        - id
        - tournament
        - name
      properties:
        id:
          type: string
          example: ufc-300
        tournament:
          type: string
          example: ufc
        name:
          type: string
          example: UFC 300
        date:
          type: string
          format: date
          example: '2025-04-13'
        location:
          type: string
          example: Las Vegas, Nevada
        venue:
          type: string
          example: T-Mobile Arena
        status:
          type: string
          enum: [upcoming, completed, cancelled]
          example: upcoming
        created_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
        updated_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
    EventInput:
      type: object
      required:
        - id
        - tournament
        - name
      properties:
        id:
          type: string
          example: ufc-300
        tournament:
          type: string
          example: ufc
        name:
          type: string
          example: UFC 300
        date:
          type: string
          format: date
          example: '2025-04-13'
        location:
          type: string
          example: Las Vegas, Nevada
        venue:
          type: string
          example: T-Mobile Arena
        status:
          type: string
          enum: [upcoming, completed, cancelled]
          example: upcoming
    Match:
      type: object
      required:
        - event_id
        - fighter1_name
        - fighter2_name
      properties:
        id:
          type: integer
          example: 1
        event_id:
          type: string
          example: ufc-300
        category:
          type: string
          example: main_card
        fighter1_id:
          type: integer
          nullable: true
          example: 1
        fighter1_name:
          type: string
          example: Fighter One
        fighter1_country:
          type: string
          example: USA
        fighter1_flag:
          type: string
          example: 🇺🇸
        fighter2_id:
          type: integer
          nullable: true
          example: 2
        fighter2_name:
          type: string
          example: Fighter Two
        fighter2_country:
          type: string
          example: Brazil
        fighter2_flag:
          type: string
          example: 🇧🇷
        weight_class:
          type: string
          example: lightweight
        winner:
          type: integer
          description: 1 for fighter1, 2 for fighter2, null for draw or no contest
          nullable: true
          example: 1
        method:
          type: string
          example: KO/TKO
        round:
          type: integer
          example: 3
        time:
          type: string
          example: '4:35'
    MatchInput:
      type: object
      required:
        - fighter1_name
        - fighter2_name
      properties:
        category:
          type: string
          example: main_card
        fighter1_id:
          type: integer
          example: 1
        fighter1_name:
          type: string
          example: Fighter One
        fighter1_country:
          type: string
          example: USA
        fighter1_flag:
          type: string
          example: 🇺🇸
        fighter2_id:
          type: integer
          example: 2
        fighter2_name:
          type: string
          example: Fighter Two
        fighter2_country:
          type: string
          example: Brazil
        fighter2_flag:
          type: string
          example: 🇧🇷
        weight_class:
          type: string
          example: lightweight
        winner:
          type: integer
          description: 1 for fighter1, 2 for fighter2, null for draw or no contest
          nullable: true
          example: 1
        method:
          type: string
          example: KO/TKO
        round:
          type: integer
          example: 3
        time:
          type: string
          example: '4:35'
...
    Promotion:
      type: object
      required:
        - id
        - name
        - subtitle
        - theme
        - color
        - events
      properties:
        id:
          type: string
          example: ufc
        name:
          type: string
          example: UFC
        subtitle:
          type: string
          example: Ultimate Fighting Championship
        theme:
          type: string
          example: ufc-theme
        color:
          type: string
          example: '#d20a0a'
        events:
          type: array
          items:
            type: string
          example: ['ufc-309', 'ufc-308']
...
  parameters:
    AthleteId:
      name: id
      in: path
      required: true
      description: Athlete ID
      schema:
        type: integer
        example: 1
    EventId:
      name: id
      in: path
      required: true
      description: Event ID
      schema:
        type: string
        example: ufc-300
    PromotionId:
      name: id
      in: path
      required: true
      description: Promotion ID
      schema:
        type: string
        example: ufc
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Athlete not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Failed to fetch athletes